Graphics.GL.ARB.SeparateShaderObjects

Description

Extension Support

glGetARBSeparateShaderObjects :: MonadIO m => m Bool

Is the ARB_separate_shader_objects extension supported?

gl_ARB_separate_shader_objects :: Bool

Is the ARB_separate_shader_objects extension supported? Note that in the presence of multiple contexts with different capabilities, this might be wrong. Use glGetARBSeparateShaderObjects in those cases instead.

Enums

pattern GL_ACTIVE_PROGRAM :: GLenum

pattern GL_ALL_SHADER_BITS :: GLbitfield

pattern GL_FRAGMENT_SHADER_BIT :: GLbitfield

pattern GL_GEOMETRY_SHADER_BIT :: GLbitfield

pattern GL_PROGRAM_PIPELINE_BINDING :: GLenum

pattern GL_PROGRAM_SEPARABLE :: GLenum

pattern GL_TESS_CONTROL_SHADER_BIT :: GLbitfield

pattern GL_TESS_EVALUATION_SHADER_BIT :: GLbitfield

pattern GL_VERTEX_SHADER_BIT :: GLbitfield

Functions

glActiveShaderProgram

Arguments

:: MonadIO m

=> GLuint
pipeline.

-> GLuint
program.

-> m ()

Manual page for OpenGL 4.x.

glBindProgramPipeline

Arguments

:: MonadIO m

=> GLuint
pipeline.

-> m ()

Manual page for OpenGL 4.x.

glCreateShaderProgramv

Arguments

:: MonadIO m

=> GLenum
type of type ShaderType.

-> GLsizei
count.

-> Ptr (Ptr GLchar)
strings pointing to count elements of type Ptr GLchar.

-> m GLuint

Manual page for OpenGL 4.x.

glDeleteProgramPipelines

Arguments

:: MonadIO m

=> GLsizei
n.

-> Ptr GLuint
pipelines pointing to n elements of type GLuint.

-> m ()

Manual page for OpenGL 4.x.

glGenProgramPipelines

Arguments

:: MonadIO m

=> GLsizei
n.

-> Ptr GLuint
pipelines pointing to n elements of type GLuint.

-> m ()

Manual page for OpenGL 4.x.

glGetProgramPipelineInfoLog

Arguments

:: MonadIO m

=> GLuint
pipeline.

-> GLsizei
bufSize.

-> Ptr GLsizei
length pointing to 1 element of type GLsizei.

-> Ptr GLchar
infoLog pointing to bufSize elements of type GLchar.

-> m ()

Manual page for OpenGL 4.x.

glGetProgramPipelineiv

Arguments

:: MonadIO m

=> GLuint
pipeline.

-> GLenum
pname of type PipelineParameterName.

-> Ptr GLint
params pointing to COMPSIZE(pname) elements of type GLint.

-> m ()

Manual page for OpenGL 4.x.

glIsProgramPipeline

Arguments

:: MonadIO m

=> GLuint
pipeline.

-> m GLboolean
of type Boolean.

Manual page for OpenGL 4.x.

glProgramParameteri

Arguments

:: MonadIO m

=> GLuint
program.

-> GLenum
pname of type ProgramParameterPName.

-> GLint
value.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform1d

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLdouble
v0.

-> m ()

glProgramUniform1dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLdouble
value pointing to count elements of type GLdouble.

-> m ()

glProgramUniform1f

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLfloat
v0.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform1fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLfloat
value pointing to count elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform1i

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLint
v0.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform1iv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLint
value pointing to count elements of type GLint.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform1ui

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLuint
v0.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform1uiv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLuint
value pointing to count elements of type GLuint.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform2d

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLdouble
v0.

-> GLdouble
v1.

-> m ()

glProgramUniform2dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLdouble
value pointing to count*2 elements of type GLdouble.

-> m ()

glProgramUniform2f

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLfloat
v0.

-> GLfloat
v1.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform2fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLfloat
value pointing to count*2 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform2i

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLint
v0.

-> GLint
v1.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform2iv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLint
value pointing to count*2 elements of type GLint.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform2ui

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLuint
v0.

-> GLuint
v1.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform2uiv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLuint
value pointing to count*2 elements of type GLuint.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform3d

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLdouble
v0.

-> GLdouble
v1.

-> GLdouble
v2.

-> m ()

glProgramUniform3dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLdouble
value pointing to count*3 elements of type GLdouble.

-> m ()

glProgramUniform3f

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLfloat
v0.

-> GLfloat
v1.

-> GLfloat
v2.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform3fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLfloat
value pointing to count*3 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform3i

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLint
v0.

-> GLint
v1.

-> GLint
v2.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform3iv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLint
value pointing to count*3 elements of type GLint.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform3ui

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLuint
v0.

-> GLuint
v1.

-> GLuint
v2.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform3uiv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLuint
value pointing to count*3 elements of type GLuint.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform4d

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLdouble
v0.

-> GLdouble
v1.

-> GLdouble
v2.

-> GLdouble
v3.

-> m ()

glProgramUniform4dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLdouble
value pointing to count*4 elements of type GLdouble.

-> m ()

glProgramUniform4f

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLfloat
v0.

-> GLfloat
v1.

-> GLfloat
v2.

-> GLfloat
v3.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform4fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLfloat
value pointing to count*4 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform4i

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLint
v0.

-> GLint
v1.

-> GLint
v2.

-> GLint
v3.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform4iv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLint
value pointing to count*4 elements of type GLint.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform4ui

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLuint
v0.

-> GLuint
v1.

-> GLuint
v2.

-> GLuint
v3.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniform4uiv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> Ptr GLuint
value pointing to count*4 elements of type GLuint.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniformMatrix2dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLdouble
value pointing to count*4 elements of type GLdouble.

-> m ()

glProgramUniformMatrix2fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLfloat
value pointing to count*4 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniformMatrix2x3dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLdouble
value pointing to count*6 elements of type GLdouble.

-> m ()

glProgramUniformMatrix2x3fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLfloat
value pointing to count*6 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniformMatrix2x4dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLdouble
value pointing to count*8 elements of type GLdouble.

-> m ()

glProgramUniformMatrix2x4fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLfloat
value pointing to count*8 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniformMatrix3dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLdouble
value pointing to count*9 elements of type GLdouble.

-> m ()

glProgramUniformMatrix3fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLfloat
value pointing to count*9 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniformMatrix3x2dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLdouble
value pointing to count*6 elements of type GLdouble.

-> m ()

glProgramUniformMatrix3x2fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLfloat
value pointing to count*6 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniformMatrix3x4dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLdouble
value pointing to count*12 elements of type GLdouble.

-> m ()

glProgramUniformMatrix3x4fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLfloat
value pointing to count*12 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniformMatrix4dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLdouble
value pointing to count*16 elements of type GLdouble.

-> m ()

glProgramUniformMatrix4fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLfloat
value pointing to count*16 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniformMatrix4x2dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLdouble
value pointing to count*8 elements of type GLdouble.

-> m ()

glProgramUniformMatrix4x2fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLfloat
value pointing to count*8 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glProgramUniformMatrix4x3dv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLdouble
value pointing to count*12 elements of type GLdouble.

-> m ()

glProgramUniformMatrix4x3fv

Arguments

:: MonadIO m

=> GLuint
program.

-> GLint
location.

-> GLsizei
count.

-> GLboolean
transpose of type Boolean.

-> Ptr GLfloat
value pointing to count*12 elements of type GLfloat.

-> m ()

Manual page for OpenGL 4.x.

glUseProgramStages

Arguments

:: MonadIO m

=> GLuint
pipeline.

-> GLbitfield
stages of type UseProgramStageMask.

-> GLuint
program.

-> m ()

Manual page for OpenGL 4.x.

glValidateProgramPipeline

Arguments

:: MonadIO m

=> GLuint
pipeline.

-> m ()

Manual page for OpenGL 4.x.