Graphics.GL.ARB.ShadingLanguageInclude

Description

Extension Support

glGetARBShadingLanguageInclude :: MonadIO m => m Bool

Is the ARB_shading_language_include extension supported?

gl_ARB_shading_language_include :: Bool

Is the ARB_shading_language_include extension supported? Note that in the presence of multiple contexts with different capabilities, this might be wrong. Use glGetARBShadingLanguageInclude in those cases instead.

Enums

pattern GL_NAMED_STRING_LENGTH_ARB :: GLenum

pattern GL_NAMED_STRING_TYPE_ARB :: GLenum

pattern GL_SHADER_INCLUDE_ARB :: GLenum

Functions

glCompileShaderIncludeARB

Arguments

:: MonadIO m

=> GLuint
shader.

-> GLsizei
count.

-> Ptr (Ptr GLchar)
path pointing to count elements of type Ptr GLchar.

-> Ptr GLint
length pointing to count elements of type GLint.

-> m ()

glDeleteNamedStringARB

Arguments

:: MonadIO m

=> GLint
namelen.

-> Ptr GLchar
name pointing to namelen elements of type GLchar.

-> m ()

glGetNamedStringARB

Arguments

:: MonadIO m

=> GLint
namelen.

-> Ptr GLchar
name pointing to namelen elements of type GLchar.

-> GLsizei
bufSize.

-> Ptr GLint
stringlen pointing to 1 element of type GLint.

-> Ptr GLchar
string pointing to bufSize elements of type GLchar.

-> m ()

glGetNamedStringivARB

Arguments

:: MonadIO m

=> GLint
namelen.

-> Ptr GLchar
name pointing to namelen elements of type GLchar.

-> GLenum
pname.

-> Ptr GLint
params pointing to COMPSIZE(pname) elements of type GLint.

-> m ()

glIsNamedStringARB

Arguments

:: MonadIO m

=> GLint
namelen.

-> Ptr GLchar
name pointing to namelen elements of type GLchar.

-> m GLboolean
of type Boolean.

glNamedStringARB

Arguments

:: MonadIO m

=> GLenum
type.

-> GLint
namelen.

-> Ptr GLchar
name pointing to namelen elements of type GLchar.

-> GLint
stringlen.

-> Ptr GLchar
string pointing to stringlen elements of type GLchar.

-> m ()