Graphics.GL.EXT.TransformFeedback

Description

Extension Support

glGetEXTTransformFeedback :: MonadIO m => m Bool

Is the EXT_transform_feedback extension supported?

gl_EXT_transform_feedback :: Bool

Is the EXT_transform_feedback extension supported? Note that in the presence of multiple contexts with different capabilities, this might be wrong. Use glGetEXTTransformFeedback in those cases instead.

Enums

pattern GL_INTERLEAVED_ATTRIBS_EXT :: GLenum

pattern GL_MAX_TRANSFORM_FEEDBACK_INTERLEAVED_COMPONENTS_EXT :: GLenum

pattern GL_MAX_TRANSFORM_FEEDBACK_SEPARATE_ATTRIBS_EXT :: GLenum

pattern GL_MAX_TRANSFORM_FEEDBACK_SEPARATE_COMPONENTS_EXT :: GLenum

pattern GL_PRIMITIVES_GENERATED_EXT :: GLenum

pattern GL_RASTERIZER_DISCARD_EXT :: GLenum

pattern GL_SEPARATE_ATTRIBS_EXT :: GLenum

pattern GL_TRANSFORM_FEEDBACK_BUFFER_BINDING_EXT :: GLenum

pattern GL_TRANSFORM_FEEDBACK_BUFFER_EXT :: GLenum

pattern GL_TRANSFORM_FEEDBACK_BUFFER_MODE_EXT :: GLenum

pattern GL_TRANSFORM_FEEDBACK_BUFFER_SIZE_EXT :: GLenum

pattern GL_TRANSFORM_FEEDBACK_BUFFER_START_EXT :: GLenum

pattern GL_TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN_EXT :: GLenum

pattern GL_TRANSFORM_FEEDBACK_VARYINGS_EXT :: GLenum

pattern GL_TRANSFORM_FEEDBACK_VARYING_MAX_LENGTH_EXT :: GLenum

Functions

glBeginTransformFeedbackEXT

Arguments

:: MonadIO m

=> GLenum
primitiveMode of type PrimitiveType.

-> m ()

This command is an alias for glBeginTransformFeedback.

glBindBufferBaseEXT

Arguments

:: MonadIO m

=> GLenum
target of type BufferTargetARB.

-> GLuint
index.

-> GLuint
buffer.

-> m ()

This command is an alias for glBindBufferBase.

glBindBufferOffsetEXT

Arguments

:: MonadIO m

=> GLenum
target of type BufferTargetARB.

-> GLuint
index.

-> GLuint
buffer.

-> GLintptr
offset of type BufferOffset.

-> m ()

glBindBufferRangeEXT

Arguments

:: MonadIO m

=> GLenum
target of type BufferTargetARB.

-> GLuint
index.

-> GLuint
buffer.

-> GLintptr
offset of type BufferOffset.

-> GLsizeiptr
size of type BufferSize.

-> m ()

This command is an alias for glBindBufferRange.

glEndTransformFeedbackEXT :: MonadIO m => m ()

This command is an alias for glEndTransformFeedback.

glGetTransformFeedbackVaryingEXT

Arguments

:: MonadIO m

=> GLuint
program.

-> GLuint
index.

-> GLsizei
bufSize.

-> Ptr GLsizei
length pointing to 1 element of type GLsizei.

-> Ptr GLsizei
size pointing to 1 element of type GLsizei.

-> Ptr GLenum
type pointing to 1 element of type GLenum.

-> Ptr GLchar
name pointing to bufSize elements of type GLchar.

-> m ()

This command is an alias for glGetTransformFeedbackVarying.

glTransformFeedbackVaryingsEXT

Arguments

:: MonadIO m

=> GLuint
program.

-> GLsizei
count.

-> Ptr (Ptr GLchar)
varyings pointing to count elements of type Ptr GLchar.

-> GLenum
bufferMode.

-> m ()

This command is an alias for glTransformFeedbackVaryings.