Graphics.Rendering.OpenGL.Raw.ARB.ShaderObjects

Description

All raw functions, tokens and types from the ARB_shader_objects extension, see http://www.opengl.org/registry/specs/ARB/shader_objects.txt.

Functions

glDeleteObject :: GLhandle -> IO ()

glGetHandle :: GLenum -> IO GLhandle

glDetachObject :: GLhandle -> GLhandle -> IO ()

glCreateShaderObject :: GLenum -> IO GLhandle

glShaderSource :: GLuint -> GLsizei -> Ptr (Ptr GLchar) -> Ptr GLint -> IO ()

glCompileShader :: GLuint -> IO ()

glCreateProgramObject :: IO GLhandle

glAttachObject :: GLhandle -> GLhandle -> IO ()

glLinkProgram :: GLuint -> IO ()

glUseProgramObject :: GLhandle -> IO ()

glValidateProgram :: GLuint -> IO ()

glUniform1f :: GLint -> GLfloat -> IO ()

glUniform2f :: GLint -> GLfloat -> GLfloat -> IO ()

glUniform3f :: GLint -> GLfloat -> GLfloat -> GLfloat -> IO ()

glUniform4f :: GLint -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> IO ()

glUniform1i :: GLint -> GLint -> IO ()

glUniform2i :: GLint -> GLint -> GLint -> IO ()

glUniform3i :: GLint -> GLint -> GLint -> GLint -> IO ()

glUniform4i :: GLint -> GLint -> GLint -> GLint -> GLint -> IO ()

glUniform1fv :: GLint -> GLsizei -> Ptr GLfloat -> IO ()

glUniform2fv :: GLint -> GLsizei -> Ptr GLfloat -> IO ()

glUniform3fv :: GLint -> GLsizei -> Ptr GLfloat -> IO ()

glUniform4fv :: GLint -> GLsizei -> Ptr GLfloat -> IO ()

glUniform1iv :: GLint -> GLsizei -> Ptr GLint -> IO ()

glUniform2iv :: GLint -> GLsizei -> Ptr GLint -> IO ()

glUniform3iv :: GLint -> GLsizei -> Ptr GLint -> IO ()

glUniform4iv :: GLint -> GLsizei -> Ptr GLint -> IO ()

glUniformMatrix2fv :: GLint -> GLsizei -> GLboolean -> Ptr GLfloat -> IO ()

glUniformMatrix3fv :: GLint -> GLsizei -> GLboolean -> Ptr GLfloat -> IO ()

glUniformMatrix4fv :: GLint -> GLsizei -> GLboolean -> Ptr GLfloat -> IO ()

glGetObjectParameterfv :: GLhandle -> GLenum -> Ptr GLfloat -> IO ()

glGetObjectParameteriv :: GLhandle -> GLenum -> Ptr GLint -> IO ()

glGetInfoLog :: GLhandle -> GLsizei -> Ptr GLsizei -> Ptr GLchar -> IO ()

glGetAttachedObjects :: GLhandle -> GLsizei -> Ptr GLsizei -> Ptr GLhandle -> IO ()

glGetUniformLocation :: GLuint -> Ptr GLchar -> IO GLint

glGetActiveUniform :: GLuint -> GLuint -> GLsizei -> Ptr GLsizei -> Ptr GLint -> Ptr GLenum -> Ptr GLchar -> IO ()

glGetUniformfv :: GLuint -> GLint -> Ptr GLfloat -> IO ()

glGetUniformiv :: GLuint -> GLint -> Ptr GLint -> IO ()

glGetShaderSource :: GLuint -> GLsizei -> Ptr GLsizei -> Ptr GLchar -> IO ()

Tokens

gl_PROGRAM_OBJECT :: GLenum

gl_OBJECT_TYPE_ARB :: GLenum

Note: We use the ARB suffix here, because the name clashes with the ARB_sync extension.

gl_OBJECT_SUBTYPE :: GLenum

gl_OBJECT_DELETE_STATUS :: GLenum

gl_OBJECT_COMPILE_STATUS :: GLenum

gl_OBJECT_LINK_STATUS :: GLenum

gl_OBJECT_VALIDATE_STATUS :: GLenum

gl_OBJECT_INFO_LOG_LENGTH :: GLenum

gl_OBJECT_ATTACHED_OBJECTS :: GLenum

gl_OBJECT_ACTIVE_UNIFORMS :: GLenum

gl_OBJECT_ACTIVE_UNIFORM_MAX_LENGTH :: GLenum

gl_OBJECT_SHADER_SOURCE_LENGTH :: GLenum

gl_SHADER_OBJECT :: GLenum

gl_FLOAT :: GLenum

gl_FLOAT_VEC2 :: GLenum

gl_FLOAT_VEC3 :: GLenum

gl_FLOAT_VEC4 :: GLenum

gl_INT :: GLenum

gl_INT_VEC2 :: GLenum

gl_INT_VEC3 :: GLenum

gl_INT_VEC4 :: GLenum

gl_BOOL :: GLenum

gl_BOOL_VEC2 :: GLenum

gl_BOOL_VEC3 :: GLenum

gl_BOOL_VEC4 :: GLenum

gl_FLOAT_MAT2 :: GLenum

gl_FLOAT_MAT3 :: GLenum

gl_FLOAT_MAT4 :: GLenum

gl_SAMPLER_1D :: GLenum

gl_SAMPLER_2D :: GLenum

gl_SAMPLER_3D :: GLenum

gl_SAMPLER_CUBE :: GLenum

gl_SAMPLER_1D_SHADOW :: GLenum

gl_SAMPLER_2D_SHADOW :: GLenum

gl_SAMPLER_2D_RECT :: GLenum

gl_SAMPLER_2D_RECT_SHADOW :: GLenum

Types

type GLhandle = CUInt