Graphics.Rendering.OpenGL.Raw.NV.VertexProgram

Description

All raw functions and tokens from the NV_vertex_program extension, see http://www.opengl.org/registry/specs/NV/vertex_program.txt.

Functions

glBindProgramNV :: GLenum -> GLuint -> IO ()

glDeleteProgramsNV :: GLsizei -> Ptr GLuint -> IO ()

glExecuteProgram :: GLenum -> GLuint -> Ptr GLfloat -> IO ()

glGenProgramsNV :: GLsizei -> Ptr GLuint -> IO ()

glAreProgramsResident :: GLsizei -> Ptr GLuint -> Ptr GLboolean -> IO GLboolean

glRequestResidentPrograms :: GLsizei -> Ptr GLuint -> IO ()

glGetProgramParameterfv :: GLenum -> GLuint -> GLenum -> Ptr GLfloat -> IO ()

glGetProgramParameterdv :: GLenum -> GLuint -> GLenum -> Ptr GLdouble -> IO ()

glGetProgramivNV :: GLuint -> GLenum -> Ptr GLint -> IO ()

glGetProgramStringNV :: GLuint -> GLenum -> Ptr GLubyte -> IO ()

glGetTrackMatrixiv :: GLenum -> GLuint -> GLenum -> Ptr GLint -> IO ()

glGetVertexAttribdvNV :: GLuint -> GLenum -> Ptr GLdouble -> IO ()

glGetVertexAttribfvNV :: GLuint -> GLenum -> Ptr GLfloat -> IO ()

glGetVertexAttribivNV :: GLuint -> GLenum -> Ptr GLint -> IO ()

glGetVertexAttribPointervNV :: GLuint -> GLenum -> Ptr (Ptr a) -> IO ()

glIsProgramNV :: GLuint -> IO GLboolean

glLoadProgram :: GLenum -> GLuint -> GLsizei -> Ptr GLubyte -> IO ()

glProgramParameter4f :: GLenum -> GLuint -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> IO ()

glProgramParameter4d :: GLenum -> GLuint -> GLdouble -> GLdouble -> GLdouble -> GLdouble -> IO ()

glProgramParameter4dv :: GLenum -> GLuint -> Ptr GLdouble -> IO ()

glProgramParameter4fv :: GLenum -> GLuint -> Ptr GLfloat -> IO ()

glProgramParameters4dv :: GLenum -> GLuint -> GLuint -> Ptr GLdouble -> IO ()

glProgramParameters4fv :: GLenum -> GLuint -> GLuint -> Ptr GLfloat -> IO ()

glTrackMatrix :: GLenum -> GLuint -> GLenum -> GLenum -> IO ()

glVertexAttribPointerNV :: GLuint -> GLint -> GLenum -> GLsizei -> Ptr a -> IO ()

glVertexAttrib1sNV :: GLuint -> GLshort -> IO ()

glVertexAttrib1fNV :: GLuint -> GLfloat -> IO ()

glVertexAttrib1dNV :: GLuint -> GLdouble -> IO ()

glVertexAttrib2sNV :: GLuint -> GLshort -> GLshort -> IO ()

glVertexAttrib2fNV :: GLuint -> GLfloat -> GLfloat -> IO ()

glVertexAttrib2dNV :: GLuint -> GLdouble -> GLdouble -> IO ()

glVertexAttrib3sNV :: GLuint -> GLshort -> GLshort -> GLshort -> IO ()

glVertexAttrib3fNV :: GLuint -> GLfloat -> GLfloat -> GLfloat -> IO ()

glVertexAttrib3dNV :: GLuint -> GLdouble -> GLdouble -> GLdouble -> IO ()

glVertexAttrib4sNV :: GLuint -> GLshort -> GLshort -> GLshort -> GLshort -> IO ()

glVertexAttrib4fNV :: GLuint -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> IO ()

glVertexAttrib4dNV :: GLuint -> GLdouble -> GLdouble -> GLdouble -> GLdouble -> IO ()

glVertexAttrib4ub :: GLuint -> GLubyte -> GLubyte -> GLubyte -> GLubyte -> IO ()

glVertexAttrib1svNV :: GLuint -> Ptr GLshort -> IO ()

glVertexAttrib1fvNV :: GLuint -> Ptr GLfloat -> IO ()

glVertexAttrib1dvNV :: GLuint -> Ptr GLdouble -> IO ()

glVertexAttrib2svNV :: GLuint -> Ptr GLshort -> IO ()

glVertexAttrib2fvNV :: GLuint -> Ptr GLfloat -> IO ()

glVertexAttrib2dvNV :: GLuint -> Ptr GLdouble -> IO ()

glVertexAttrib3svNV :: GLuint -> Ptr GLshort -> IO ()

glVertexAttrib3fvNV :: GLuint -> Ptr GLfloat -> IO ()

glVertexAttrib3dvNV :: GLuint -> Ptr GLdouble -> IO ()

glVertexAttrib4svNV :: GLuint -> Ptr GLshort -> IO ()

glVertexAttrib4fvNV :: GLuint -> Ptr GLfloat -> IO ()

glVertexAttrib4dvNV :: GLuint -> Ptr GLdouble -> IO ()

glVertexAttrib4ubvNV :: GLuint -> Ptr GLubyte -> IO ()

glVertexAttribs1sv :: GLuint -> GLsizei -> Ptr GLshort -> IO ()

glVertexAttribs1fv :: GLuint -> GLsizei -> Ptr GLfloat -> IO ()

glVertexAttribs1dv :: GLuint -> GLsizei -> Ptr GLdouble -> IO ()

glVertexAttribs2sv :: GLuint -> GLsizei -> Ptr GLshort -> IO ()

glVertexAttribs2fv :: GLuint -> GLsizei -> Ptr GLfloat -> IO ()

glVertexAttribs2dv :: GLuint -> GLsizei -> Ptr GLdouble -> IO ()

glVertexAttribs3sv :: GLuint -> GLsizei -> Ptr GLshort -> IO ()

glVertexAttribs3fv :: GLuint -> GLsizei -> Ptr GLfloat -> IO ()

glVertexAttribs3dv :: GLuint -> GLsizei -> Ptr GLdouble -> IO ()

glVertexAttribs4sv :: GLuint -> GLsizei -> Ptr GLshort -> IO ()

glVertexAttribs4fv :: GLuint -> GLsizei -> Ptr GLfloat -> IO ()

glVertexAttribs4dv :: GLuint -> GLsizei -> Ptr GLdouble -> IO ()

glVertexAttribs4ubv :: GLuint -> GLsizei -> Ptr GLubyte -> IO ()

Tokens

gl_VERTEX_PROGRAM :: GLenum

gl_VERTEX_PROGRAM_POINT_SIZE :: GLenum

gl_VERTEX_PROGRAM_TWO_SIDE :: GLenum

gl_VERTEX_STATE_PROGRAM :: GLenum

gl_ATTRIB_ARRAY_SIZE :: GLenum

gl_ATTRIB_ARRAY_STRIDE :: GLenum

gl_ATTRIB_ARRAY_TYPE :: GLenum

gl_CURRENT_ATTRIB :: GLenum

gl_PROGRAM_PARAMETER :: GLenum

gl_ATTRIB_ARRAY_POINTER :: GLenum

gl_PROGRAM_TARGET :: GLenum

gl_PROGRAM_LENGTH :: GLenum

gl_PROGRAM_RESIDENT :: GLenum

gl_PROGRAM_STRING :: GLenum

gl_TRACK_MATRIX :: GLenum

gl_TRACK_MATRIX_TRANSFORM :: GLenum

gl_MAX_TRACK_MATRIX_STACK_DEPTH :: GLenum

gl_MAX_TRACK_MATRICES :: GLenum

gl_CURRENT_MATRIX_STACK_DEPTH :: GLenum

gl_CURRENT_MATRIX :: GLenum

gl_VERTEX_PROGRAM_BINDING :: GLenum

gl_PROGRAM_ERROR_POSITION :: GLenum

gl_MODELVIEW_PROJECTION :: GLenum

gl_MATRIX0_NV :: GLenum

Note: We use the NV suffix here, because the name clashes with the ARB_vertex_program.

gl_MATRIX1_NV :: GLenum

Note: We use the NV suffix here, because the name clashes with the ARB_vertex_program.

gl_MATRIX2_NV :: GLenum

Note: We use the NV suffix here, because the name clashes with the ARB_vertex_program.

gl_MATRIX3_NV :: GLenum

Note: We use the NV suffix here, because the name clashes with the ARB_vertex_program.

gl_MATRIX4_NV :: GLenum

Note: We use the NV suffix here, because the name clashes with the ARB_vertex_program.

gl_MATRIX5_NV :: GLenum

Note: We use the NV suffix here, because the name clashes with the ARB_vertex_program.

gl_MATRIX6_NV :: GLenum

Note: We use the NV suffix here, because the name clashes with the ARB_vertex_program.

gl_MATRIX7_NV :: GLenum

Note: We use the NV suffix here, because the name clashes with the ARB_vertex_program.

gl_IDENTITY :: GLenum

gl_INVERSE :: GLenum

gl_TRANSPOSE :: GLenum

gl_INVERSE_TRANSPOSE :: GLenum

gl_VERTEX_ATTRIB_ARRAY0 :: GLenum

gl_VERTEX_ATTRIB_ARRAY1 :: GLenum

gl_VERTEX_ATTRIB_ARRAY2 :: GLenum

gl_VERTEX_ATTRIB_ARRAY3 :: GLenum

gl_VERTEX_ATTRIB_ARRAY4 :: GLenum

gl_VERTEX_ATTRIB_ARRAY5 :: GLenum

gl_VERTEX_ATTRIB_ARRAY6 :: GLenum

gl_VERTEX_ATTRIB_ARRAY7 :: GLenum

gl_VERTEX_ATTRIB_ARRAY8 :: GLenum

gl_VERTEX_ATTRIB_ARRAY9 :: GLenum

gl_VERTEX_ATTRIB_ARRAY10 :: GLenum

gl_VERTEX_ATTRIB_ARRAY11 :: GLenum

gl_VERTEX_ATTRIB_ARRAY12 :: GLenum

gl_VERTEX_ATTRIB_ARRAY13 :: GLenum

gl_VERTEX_ATTRIB_ARRAY14 :: GLenum

gl_VERTEX_ATTRIB_ARRAY15 :: GLenum

gl_MAP1_VERTEX_ATTRIB0_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB1_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB2_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB3_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB4_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB5_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB6_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB7_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB8_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB9_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB10_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB11_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB12_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB13_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB14_4 :: GLenum

gl_MAP1_VERTEX_ATTRIB15_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB0_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB1_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB2_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB3_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB4_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB5_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB6_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB7_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB8_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB9_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB10_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB11_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB12_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB13_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB14_4 :: GLenum

gl_MAP2_VERTEX_ATTRIB15_4 :: GLenum