Graphics.Rendering.OpenGL.Raw.EXT.BindableUniform

Description

All raw functions and tokens from the EXT_bindable_uniform extension not already in the OpenGL 3.1 core, see http://www.opengl.org/registry/specs/EXT/bindable_uniform.txt.

Functions

glUniformBuffer :: GLuint -> GLint -> GLuint -> IO ()

glGetUniformBufferSize :: GLuint -> GLint -> IO GLint

glGetUniformOffset :: GLuint -> GLint -> IO GLintptr

Tokens

gl_MAX_VERTEX_BINDABLE_UNIFORMS :: GLenum

gl_MAX_FRAGMENT_BINDABLE_UNIFORMS :: GLenum

gl_MAX_GEOMETRY_BINDABLE_UNIFORMS :: GLenum

gl_MAX_BINDABLE_UNIFORM_SIZE :: GLenum

gl_UNIFORM_BUFFER_BINDING_EXT :: GLenum

NOTE: The OpenGL 3.1 core has the same token with a different value, so we use a suffix here.

gl_UNIFORM_BUFFER_EXT :: GLenum

NOTE: The OpenGL 3.1 core has the same token with a different value, so we use a suffix here.