Graphics.Rendering.OpenGL.Raw.EXT.Multisample

Description

All raw functions and tokens from the EXT_multisample extension, see http://www.opengl.org/registry/specs/EXT/multisample.txt.

Functions

glSampleMask :: GLclampf -> GLboolean -> IO ()

glSamplePattern :: GLenum -> IO ()

Tokens

gl_MULTISAMPLE :: GLenum

gl_SAMPLE_ALPHA_TO_MASK :: GLenum

gl_SAMPLE_ALPHA_TO_ONE :: GLenum

gl_SAMPLE_MASK_EXT :: GLenum

Note: We use the EXT suffix here, because the name clashes with the ARB_texture_multisample extension.

gl_MULTISAMPLE_BIT :: GLenum

gl_1PASS :: GLenum

gl_2PASS_0 :: GLenum

gl_2PASS_1 :: GLenum

gl_4PASS_0 :: GLenum

gl_4PASS_1 :: GLenum

gl_4PASS_2 :: GLenum

gl_4PASS_3 :: GLenum

gl_SAMPLE_BUFFERS :: GLenum

gl_SAMPLES :: GLenum

gl_SAMPLE_MASK_VALUE_EXT :: GLenum

Note: We use the EXT suffix here, because the name clashes with the ARB_texture_multisample extension.

gl_SAMPLE_MASK_INVERT :: GLenum

gl_SAMPLE_PATTERN :: GLenum