Graphics.Rendering.OpenGL.Raw.GetProcAddress

Description

This module offers a portable way to retrieve OpenGL extension entries, providing a portability layer upon platform-specific mechanisms like glXGetProcAddress, wglGetProcAddress or NSAddressOfSymbol.

Documentation

getProcAddress :: String -> IO (FunPtr a)

Retrieve an OpenGL extension entry by name. Returns nullFunPtr when no extension entry with the given name was found.

getProcAddressWithSuffixes :: String -> [String] -> IO (FunPtr a)

Retrieve an OpenGL extension entry by name, trying a list of name suffixes in the given order. Returns nullFunPtr when no extension entry with the given name plus any of the suffixes was found.